% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{as.nodes}
\alias{as.nodes}
\alias{as.nodes.data.frame}
\alias{as.nodes.list}
\alias{as.nodes.numeric}
\alias{as.nodes.probField}
\alias{as.nodes.probFieldNodeList}
\title{Convert data to nodes}
\usage{
as.nodes(x, ...)

\method{as.nodes}{data.frame}(x, response = NULL, ...)

\method{as.nodes}{numeric}(x, ...)

\method{as.nodes}{list}(x, ...)

\method{as.nodes}{probFieldNodeList}(x, ...)

\method{as.nodes}{probField}(x, ...)
}
\arguments{
\item{x}{A data.frame or vector}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{response}{Which column of a data frame to be considered the random variable}
}
\value{
A \code{probFieldNodeList}
}
\description{
Generic function to convert arbitrary data types into \code{probFieldNodeList} objects
}
\details{
If data frame is supplied each row is converted into a node with one column optionally representing
the response variable, if a numeric vector is supplied each element is considered a guide variable
for a new node. Lists are cast as class \code{probFieldNodeList}, \code{ProbField}s have their
survey nodes returned, and \code{probFieldNodeLists} are returned unmodified.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Convert a data.frame to a \code{probFieldNodeList}

\item \code{numeric}: 

\item \code{list}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

