% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{breakTies}
\alias{breakTies}
\title{Tie breaking}
\usage{
breakTies(x)
}
\arguments{
\item{x}{A vector of observations}
}
\value{
A vector of observations with ties broken
}
\description{
The functions takes a vector of observations, generates a sample from 1 to the number of
observations without replacement, and multiplies that by the smallest positive double value
\code{.Machine$double.xmin} and adds that to the original observations. The procedure should
work in almost all cases except when values are measured with precision of 10e-308, in
which case ties should be unlikely at the outset.
}

