% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{calcCDF}
\alias{calcCDF}
\alias{calcCDF.list}
\alias{calcCDF.probFieldNode}
\alias{calcCDF.probFieldNodeList}
\title{Generate the CDF, Step probabilities, and Redundancy}
\usage{
calcCDF(x, ...)

\method{calcCDF}{probFieldNodeList}(x, kernel, distance,
  redundancyCorrected = TRUE, ...)

\method{calcCDF}{list}(x, ...)

\method{calcCDF}{probFieldNode}(x, nodeList, kernel, distance,
  redundancyCorrected = TRUE, ...)
}
\arguments{
\item{x}{Either A \code{probFieldNode} or a \code{probFieldNodeList}}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{kernel}{The kernel function of interest}

\item{distance}{The distance function of interest}

\item{redundancyCorrected}{Whether or not to adjust for inter node redundancy}

\item{nodeList}{A \code{probFieldNodeList} with nodes containing precalculated redundancy}
}
\value{
An object with the same type as x with step probabilities, CDF(s), and redundancy
attached
}
\description{
Generates the CDF, step probabilities, and redundancy for nodes either for a
list of nodes, or for one node relative to another list of nodes
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNodeList}: Generate CDFs, step probabilities, and redundancy for a
\code{probFieldNodeList}

\item \code{list}: 

\item \code{probFieldNode}: Calculate the CDF, step probabilities, and redundancy for a \code{probFieldNode}
given a \code{probFieldNodeList}
}}

