% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concepts.R
\name{concepts}
\alias{concepts}
\title{Ogmap Concepts}
\description{
The easiest way to figure out a new package is to understand the key concepts and the problem the
package was designed to solve. In this case the ogmap package was created out of the desire to have a way
to estimate full probability distributions for a random variable of interest without resorting
to the standard distributional assumptions used in the standard parametric toolkit

To better understand how ogmap works it may be worthwhile to read the following conceptual
documentation as needed.
\itemize{
  \item{Probability Fields:}{ \link{probField-concept}}
  \item{Nodes:}{ \link{nodes-concept}}
  \item{Internode Influence:}{ \link{influence-concept}}
  \item{Inference from the Field:}{ \link{inference-concept}}
  \item{Field Correctness:}{ \link{correctness-concept}}
}
}

