% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concepts.R
\name{correctness-concept}
\alias{correctness-concept}
\title{Field Correctness}
\description{
The quality of the inference generated from a probability field is dependent on the quality (correctness)
of the distributions assigned by the field. If correct probability distributions
have been assigned then the survey observations should be uniformly distributed throughout their respective
distributions and nearby pairs of nodes should be independent of one another.

How well the field satisfies these two criteria depends on the choice of kernel and distance functions
(see \link{influence-concept}). Once the kernel and distance functions have been chosen the shape
and scale parameters that govern their behaviour can be optimized to satisfy the
two criteria above (see \link{optimizeBandwidths}). This ensures that the distributions assigned by
the probability field are as correct as possible given the survey data.

The optimization procedure hinges on knowing the probabilities that both criteria have been satisfied.
The probability that a survey node is uniformly distributed within its probability distributions is
given by the jackknife probability of that survey node. The set of these probabilities can be compared
to the identity line (x) with a Kolmogorov-Smirnov test. If nearby pairs of nodes are independent of
one-another, the difference in jackknife probabilities between each pair should be triangularly distributed,
which can be assessed with another Kolmogorov-Smirnov test against  \eqn{2x - x^2}.  The probabilities derived
from these two tests can be combined into a single optimizable quantity:.

\deqn{product * (1 - log(product))}
}

