% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{debugKernel}
\alias{debugKernel}
\title{Debug Kernel
A simple kernel for debugging}
\usage{
debugKernel()
}
\value{
A \code{probFieldKernel} wrapping a function of distance which returns 0.5 if distance is less than
1, 0.5 if distance is between .5 and 1.5 exclusive, and .25 if distance is less than 2.5 and zero for any other
input
}
\description{
Debug Kernel
A simple kernel for debugging
}

