% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{distance}
\alias{distance}
\alias{probFieldDistance}
\title{Create an inter-node distance function}
\usage{
distance(fun, ..., guides = NULL)
}
\arguments{
\item{fun}{A function of two nodes}

\item{...}{Addition scale parameters used in the calculation of distance}

\item{guides}{An optional named list of references (either named or positional) to the guide members
that are used in the distance calculation}
}
\value{
A \code{probFieldDistance} object, which is a wrapped function with attributes
representing scale parameters
}
\description{
Create a distance function to be used in calculating the distance between two
\code{probFieldNodes}
}
\details{
This function constructs a distance metric of the users choosing to be used in calculations
within and involving probability fields. Run \code{linearDistance} (without brackets) and
arguments to see a simple sample implementation. For compatibility the function must take
two arguments (each a \code{probFieldNode}) extract the guide variables and calculate a
meaningful distance metric to be used by \code{probFieldKernel} objects. For a complex example
see the code for \link{trawlDistance}.
}

