% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{earthDistance}
\alias{earthDistance}
\title{Earth Surface Distance}
\usage{
earthDistance(horizontalScale, latGuide, longGuide, earthRadius = 6367)
}
\arguments{
\item{horizontalScale}{The scaling factor to apply to the distance}

\item{latGuide}{The name of the guide parameter containing the latitudinal coordinate
within the \code{probFieldNode}s}

\item{longGuide}{The name of the guide parameter containing the longitudinal coordinate
within the \code{probFieldNode}s}

\item{earthRadius}{The radius of the earth in kilometers at the approximate latitude of interest
defaults to the radius of the earth at 55N}
}
\description{
Create \code{probFieldDistance} object to calculate the distance between two nodes
along the surface of the earth, adjusting for curvature, given lat and long coordinates.
}

