% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{evaluateNodes}
\alias{evaluateNodes}
\alias{evaluateNodes.probFieldNode}
\alias{evaluateNodes.probFieldNodeList}
\title{Add Information to a nodes from a field}
\usage{
evaluateNodes(nodes, field)

\method{evaluateNodes}{probFieldNode}(nodes, field)

\method{evaluateNodes}{probFieldNodeList}(nodes, field)
}
\arguments{
\item{nodes}{The new \code{probFieldNode} or \code{probFieldNodeList}}

\item{field}{The \code{probField} against which to evaluate the node's CDF}
}
\description{
For a node or nodes within the domain of a probability field, calculate the CDF for the new
node relative to the survey nodes of the field, as well as the field's kernel and
distance functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 
}}

