% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistic.R
\name{expectationStat}
\alias{expectationStat}
\title{Calculate Expected Values}
\usage{
expectationStat(field, nodes)
}
\arguments{
\item{field}{The \code{probField} of interest}

\item{nodes}{Either a \code{probFieldNode} or \code{probFieldNodeList} of nodes for which to calculate the
expectation}
}
\value{
A \code{probFieldStatistic} object containing the observed values from the survey nodes in the field,
a callable result function to allow statistic resampling, and the results of applying the design function
to the field and nodes provided, in this case a matrix of steps probabilities for the nodes provided and
the values from the field.
}
\description{
Generate a \code{probFieldStatistic} to provide easy resampling of expected values for one or many nodes
}

