% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptionChecking.R
\name{findNearbyPairs}
\alias{findNearbyPairs}
\title{Find Close Pairs of Nodes}
\usage{
findNearbyPairs(nodeSource, distance, nStarts = length(nodeSource))
}
\arguments{
\item{nodeSource}{An object that can be coerced to a \code{probFieldNodeList} (through
\link{as.nodes})}

\item{distance}{The distance function used to evaluate inter-node distances}

\item{nStarts}{How many randomly selected nodes to try, defaults to all possible nodes,
alternatively a vector of indices to the starting nodes.}
}
\value{
A 2-column matrix with the indices to each pair in the best path
}
\description{
Find a set of node pairing that describe a reasonable short path through the data.
Uses the nearest-neighbour solution to the travelling salesman problem using
a variety of start positions to find the shortest such path
}

