% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{findRelevantIndices}
\alias{findRelevantIndices}
\title{Find Relevant Indices}
\usage{
findRelevantIndices(nodes, surveySource, residue = 0.02, weights = NULL)
}
\arguments{
\item{nodes}{The \code{probFieldNode} or \code{probFieldNodeList} to for which to find
their influential counterparts}

\item{surveySource}{Either a \code{probField} or \code{probFieldNodeList} from which to find
the indices of relevant member nodes}

\item{residue}{The percent of the right tail of the weighted relevance distribution to discard}

\item{weights}{A vector of weights representing the relative importance of influence on
each of the nodes given to the \code{nodes} parameter}
}
\value{
A vector of indices corresponding to the positions within \code{surveySource} that
are relevant to members of \code{nodes}
}
\description{
Find the indices corresponding to a subset of nodes from a group of nodes
(either a \code{probField} or \code{probFieldNodeList}) that has a given amount of
relevance to another group of nodes. The influence may be weighted by the importance
of nodes within the other group.
}
\details{
surveySource should be a \code{probField} or \code{probFieldNodeList}, although
any object that can be coerced to a \code{probFieldNodeList} works, however doing so
is ill-advised
}

