% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{getBandwidths}
\alias{getBandwidths}
\title{Retrieve bandwidths}
\usage{
getBandwidths(field)
}
\arguments{
\item{field}{The \code{probField} of interest}
}
\value{
A named character vector of additional parameters with an additional \code{from}
attribute specifying whether the parameters belong to the kernel or distance function
}
\description{
Return all additional parameters present in the kernel and distance components of a
probability field
}

