% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getCDF}
\alias{getCDF}
\alias{getCDF.probField}
\alias{getCDF.probFieldNode}
\alias{getCDF.probFieldNodeList}
\title{Retrieve Node CDFs}
\usage{
getCDF(x, ...)

\method{getCDF}{probFieldNode}(x, ...)

\method{getCDF}{probFieldNodeList}(x, asMatrix = TRUE, ...)

\method{getCDF}{probField}(x, asMatrix = TRUE, ...)
}
\arguments{
\item{x}{A \code{probFieldNode}, \code{probFieldNodeList}, or \code{probField} from which to obtain the CDF(s)}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{asMatrix}{Logical whether the results should be returned as matrix or list when retrieving CDFs from a
collection of nodes}
}
\value{
A vector in the case x is a \code{probFieldNode} otherwise a matrix or list of such vectors at the users
discretion
}
\description{
Return the CDF for a \code{probFieldNode} or a group thereof
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

