% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{getDistanceParameters}
\alias{getDistanceParameters}
\alias{getScale}
\title{Retrieve Distance Parameters}
\usage{
getDistanceParameters(distance)

getScale(distance)
}
\arguments{
\item{distance}{A \code{probFieldDistance}}
}
\value{
A named vector containing additional parameters and their values
}
\description{
Extract the additional parameters used by a distance function
}
\details{
getScale is an alias to getDistanceParameters, the name of this function
derives from the scale parameter used in \link{linearDistance}
}

