% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getGuide}
\alias{getGuide}
\alias{getGuide.probField}
\alias{getGuide.probFieldNode}
\alias{getGuide.probFieldNodeList}
\title{getGuide}
\usage{
getGuide(x, ...)

\method{getGuide}{probFieldNode}(x, returnVector = TRUE, ...)

\method{getGuide}{probFieldNodeList}(x, ...)

\method{getGuide}{probField}(x, ...)
}
\arguments{
\item{x}{A \code{probFieldNode}, \code{probFieldNodeList}, or \code{probField}}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{returnVector}{Force the guide(s) to be returned as a vector}
}
\value{
For a \code{probFieldNode}, a vector containing the guide values for a node,
or if returnVector is set to false, potentially a one row data.frame depending on how
the node was constructed. For a collection of nodes returns a data.frame of guide variable values
}
\description{
Get the guide variable(s) for a given node or collection thereof
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

