% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getIndex}
\alias{getIndex}
\title{getIndex}
\usage{
getIndex(node, nodeListSource)
}
\arguments{
\item{node}{The \code{probFieldNode} of interest}

\item{nodeListSource}{Either a \code{probFieldNodeList} or a \code{probField}
from which to retreave a nodeList of survey nodes.}
}
\value{
A numeric index specifying the position of the given node in the list
}
\description{
Return the index identifying the position of a given node in a list of nodes
or a probability field
}

