% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{getKernelParameters}
\alias{getKernelParameters}
\alias{getShape}
\title{Kernel Parameters}
\usage{
getKernelParameters(kernel)

getShape(kernel)
}
\arguments{
\item{kernel}{The \code{probFieldKernel} of interest}
}
\value{
A named vector containing the additional shape parameters used in the kernel function
}
\description{
Return a vector of the additional shape parameters used in a kernel function
}
\details{
\code{getShape} is an alias to \code{getKernelParameters} this function's name derives
from the single shape parameter necessary for many kernel functions. See the implementation of
\link{subcauchy} for an example
}

