% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{getNode}
\alias{getNode}
\title{Retrieve a node}
\usage{
getNode(field, lookup, type = c("index", "value", "guide"),
  tolerance = 10^-5, allowMultiples = FALSE)
}
\arguments{
\item{field}{The probability field of interest}

\item{lookup}{The value be found}

\item{type}{A choice of index, value, or guide, (defaulting to index) for how to find
the node of interest}

\item{tolerance}{If matching by guide or value, what is the maximum difference between
the lookup and the node value to be considered a match}

\item{allowMultiples}{when matching by value or guide should multiple values potentially
be returned, if not the first match is returned}
}
\value{
A \code{probFieldNode} or NULL if no match is found when allowMultiples is set
to FALSE, else, a \code{probFieldNodeList} containing one or more matches
}
\description{
Function to retrieve a specific node in a field from either it's guide, value, or index
}

