% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getRedundancy}
\alias{getRedundancy}
\alias{getRedundancy.probField}
\alias{getRedundancy.probFieldNode}
\alias{getRedundancy.probFieldNodeList}
\title{Retrieve Node Redundancy}
\usage{
getRedundancy(x)

\method{getRedundancy}{probFieldNode}(x)

\method{getRedundancy}{probFieldNodeList}(x)

\method{getRedundancy}{probField}(x)
}
\arguments{
\item{x}{A \code{probFieldNode}, \code{probFieldNodeList}, or \code{probField}}
}
\value{
A numeric vector containing the redundancy of all nodes of interest
}
\description{
Return the redundancy of a survey node or a collection thereof
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

