% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getSteps}
\alias{getSteps}
\alias{getSteps.probField}
\alias{getSteps.probFieldNode}
\alias{getSteps.probFieldNodeList}
\title{Retrieve Step Probabilities}
\usage{
getSteps(x, ...)

\method{getSteps}{probFieldNode}(x, ...)

\method{getSteps}{probFieldNodeList}(x, asMatrix = TRUE, ...)

\method{getSteps}{probField}(x, asMatrix = TRUE, ...)
}
\arguments{
\item{x}{A \code{probFieldNode}, \code{probFieldNodeList}, or \code{probField} from which to obtain step probabilities}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{asMatrix}{Logical whether the results should be returned as matrix or list when retrieving steps from a
collection of nodes}
}
\value{
A vector in the case x is a \code{probFieldNode} otherwise a matrix or list of such vectors at the users
discretion
}
\description{
Return the step probabilities for a \code{probFieldNode} or a group thereof
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

