% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{getVariable}
\alias{getVariable}
\alias{getVariable.probField}
\alias{getVariable.probFieldNode}
\alias{getVariable.probFieldNodeList}
\title{getVariable}
\usage{
getVariable(x)

\method{getVariable}{probFieldNode}(x)

\method{getVariable}{probFieldNodeList}(x)

\method{getVariable}{probField}(x)
}
\arguments{
\item{x}{A \code{probFieldNode}, \code{probFieldNodeList}, or \code{probField}}

\item{...}{Additional parameters (currently only included for the reference in the generic)}
}
\value{
For a node, either a number or NA depending on whether the node contains an
observation. For a nodeList, a vector of observations
}
\description{
Return the random variable observation of a \code{probFieldNode} or collection thereof
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldNode}: 

\item \code{probFieldNodeList}: 

\item \code{probField}: 
}}

