% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concepts.R
\name{inference-concept}
\alias{inference-concept}
\title{Inference From the Field}
\description{
The point of creating a probability field is to be able to use it to perform inference. The shape this
inference procedure takes depends critically on the question being asked. All inference is made with respect
to the random variable at some set of inference nodes (see: \link{nodes-concept}). The inference can be a point
estimate, i.e. assessing the distribution of the random variable with respect to one or more nodes independently,
or the inference can be dependent on the distribution at multiple nodes (some form of weighted sum).

Since the distribution at each inference node depends on the survey nodes used to build the field, bootstrap resampling
of the survey nodes allows you to place confidence limits on the estimates obtained by most inference procedures.
To facillitate resampling based techniques we've provided the statistic class (see \link{probFieldStatistic}).
Objects of this class can be used to rapidly resample a statistic of your choice.
}

