% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concepts.R
\name{influence-concept}
\alias{influence-concept}
\title{Internode Influence - Concept}
\description{
In order to create the probability distribution at an inference node it is necessary to determine how
to weight the observations at the survey nodes. This process requires a distance function and a weighting
function (kernel), taken together they measure the influence of the survey node on the distribution at the inference
node. This formalizes the intuition that the expected value of a node should be most similar to the nodes closer to it
in guide space. The first step is to assess how close the nodes are to one another. The distance between nodes
can be simple distances like euclidean distance, more complicated like spherical surface distance, or
even non-metric distances like string-edit or Hamming distances. The result of the distance function is then passed
to a kernel function which determines the weight an observation at that distance should have on the distribution.

Serval prepared kernel and distance functions are made available by the ogmap package, although many cases may
require bespoke versions. Personal kernel and distance functions can be constructed through \link{distance}
and \link{kernel} functions.
}

