% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptionChecking.R
\name{jackknifePairProbDiff}
\alias{jackknifePairProbDiff}
\title{Jackknife pair probability difference}
\usage{
jackknifePairProbDiff(i, j, field)
}
\arguments{
\item{i}{Index of the first \code{probFieldNode} to be considered}

\item{j}{Index of the second \code{probFieldNode} to be considered}

\item{field}{the \code{probField} to be considered}
}
\description{
Calculate the jackknife probability difference between a pair
of nodes
}

