% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptionChecking.R
\name{jackknifeProb}
\alias{jackknifeProb}
\title{Jackknife probability of a node}
\usage{
jackknifeProb(x, field)
}
\arguments{
\item{x}{Either a \code{probFieldNode} or a numeric index to a node of interest
within a survey}

\item{field}{The probability field from which to obtain the survey nodes}
}
\value{
A numeric probability
}
\description{
Calculate the jackknife probability of a node
}

