% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{kernel}
\alias{kernel}
\alias{probFieldKernel}
\title{Create a kernel weighting function}
\usage{
kernel(fun, ...)
}
\arguments{
\item{fun}{A function of distance}

\item{...}{Extra shape parameters}
}
\value{
A \code{probFieldKernel} object that wraps a kernel function with an additional
attribute containing the shaping parameters, and the call used to produce the kernel function
}
\description{
Create a custom \code{probFieldKernel} object to evaluate the relevance of an
observation at a given distance
}
\details{
The function allows a user to create a custom kernel function to assign a weighting to an
observation at a given distance. The provided function needs only be a function of a number
representing a distance.
}

