% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{linearDistance}
\alias{linearDistance}
\title{Linear Distance Between Nodes}
\usage{
linearDistance(scale, guides = NULL)
}
\arguments{
\item{scale}{The scaling parameter}

\item{guides}{A vector containing the names of the guide variables used in the distance calculation,
must be provided for multidimensional guide variables}
}
\value{
A \code{probFieldDistance} object which wraps a function that extracts the guide
values from two given nodes and returns the absolute linear distance
}
\description{
Generates a \code{probFieldDistance} object that calculates the absolute linear
distance between two nodes divided by a given scale
}
\details{
This code is designed to make a \code{probFieldDistance} object that will calculate a
linear distance between the guides of two nodes. It performs pair-wise subtractions
and element-wise division, the distances in each guide variable are summed
}

