% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{logist}
\alias{logist}
\title{Logistic Kernel}
\usage{
logist(shape)
}
\arguments{
\item{shape}{The shape parameter governing the logistic kernel}
}
\value{
A \code{probFieldKernel} object which wraps a logistic kernel function with the shape
parameter \code{shape} and an additional attribute containing the value of \code{shape}
}
\description{
Generate a \code{probFieldKernel} object to evaluate the relevance of an observation at a give distance
using a logistic kernel governed by a single shape parameter
}

