% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{Create a node for a probability field}
\usage{
node(guide, ranVar = NULL, steps = NULL, ogive = NULL, redundancy = 1)
}
\arguments{
\item{guide}{An arbitrary length vector of predictor variables
(although for almost all purposes its length should match its associated field)}

\item{ranVar}{An option observation of a random variable of interest at the given guide}

\item{steps}{The step probabilities for ordered nodes in an associated field}

\item{ogive}{The cdf of step probabilities for ordered nodes in an associated field}

\item{redundancy}{The inferential redundancy of this node relative to the other nodes in
an associated field}
}
\value{
A \code{probFieldNode} object containing all the parameter values
}
\description{
Create a node for a probability field, or to generate inference from a probability field.
A node contains a guide - a vector of predictor variables - and an optional random variable
}
\details{
Users should strongly consider avoiding manually supplying steps,
ogives, and redundancy values. These values are typically calculated by
\link{calcCDF} and only make sense when in context of another set of nodes
}

