% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{nodeExpectation}
\alias{nodeConfidenceIntervals}
\alias{nodeDiversity}
\alias{nodeExpectation}
\alias{nodeVariance}
\title{Node Summary Statistics}
\usage{
nodeExpectation(node, field)

nodeVariance(node, field)

nodeDiversity(node)

nodeConfidenceIntervals(node, field, confidence = 0.95, resamples = 1000)
}
\arguments{
\item{node}{The \code{probFieldNode} of interest}

\item{field}{The \code{probField} of interest}

\item{confidence}{The level of confidence for the interval}

\item{resamples}{The number of bootstrap resamples used to calculate confidence
limits}
}
\value{
The numeric value of the statistic for the node for expectation, variance
and diversity, and a length-3 named vector with lower-limit, expectation, upper-limit
for confidence intervals
}
\description{
The expected value, variance, or diversity of a given node
}
\section{Functions}{
\itemize{
\item \code{nodeVariance}: 

\item \code{nodeDiversity}: 

\item \code{nodeConfidenceIntervals}: 
}}

