% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{nodeRelevance}
\alias{nodeRelevance}
\title{Inter Node Relevance}
\usage{
nodeRelevance(node1, node2, kernel, distance)
}
\arguments{
\item{node1}{a \code{probFieldNode} object}

\item{node2}{another \code{probFieldNode}}

\item{kernel}{the kernel function of interest}

\item{distance}{the distance function of interest}
}
\description{
Calculation to determine the inferential relevance of one node to another.
Simply evaluates the kernel function on the distance between two nodes given
a distance function.
}

