% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ogmap.R
\docType{package}
\name{ogmap}
\alias{ogmap}
\alias{ogmap-package}
\title{Probability distribution estimation dependent on covariates}
\description{
The ogmap package is designed to overcome the limitation of using only the expected value of a distribution for
inference. The package uses non-parametric weighting of survey observations to allow the entire distribution of
potential values given predictor variables to be examined.

This package was designed with funding from the Canadian government's Department of Fisheries and Oceans. This
is a port of a python package designed, written, and conceived by Geoff Evans, with the port created by Chris Hammill
}
\author{
Chris Hammill

Geoff Evans
}

