% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizeBandwidths}
\alias{optimizeBandwidths}
\title{Bandwidth Optimizer}
\usage{
optimizeBandwidths(field, pairs, method = c("optim", "cmaes"), ...,
  verbose = FALSE, returnField = TRUE)
}
\arguments{
\item{field}{The \code{probField} for which to optimize kernel and distance function parameters}

\item{pairs}{A 2 column matrix or \code{data.frame} containing integer indices to pairs of survey nodes to be
used in the optimization function}

\item{method}{The optimization algorithm to use, currently either optim (Nelder-Mead; default) or cmaes (Covariance
matrix adapting evolutionary strategy)}

\item{...}{Additional parameters to be passed to the optimization algorithm}

\item{verbose}{Whether or not to print the optimized bandwidths}

\item{returnField}{Whether or not to return the new field with optimized bandwidths or the bandwidth values}
}
\value{
If returnField is TRUE returns a \code{probField} with the optimized bandwidths, if FALSE the results of
the optimizer function (see \link{optim} and \link{cma_es} for details)
}
\description{
determine the optimum kernel and distance parameters for a \code{probField} given paired observations
using either Nelder-Mead optimization via \link{optim} or covariance matrix adapting evolutionary strategy
optimization via \link{cma_es}
}

