% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assumptionChecking.R
\name{plausibleIndependentPairs}
\alias{plausibleIndependentPairs}
\title{Assess Paired Node Independence}
\usage{
plausibleIndependentPairs(field, rawPairs)
}
\arguments{
\item{field}{The \code{probField} containing the nodes of interest}

\item{rawPairs}{A 2-column matrix or data.frame denoting the index of the members of the pairs with respect
to the original survey data}
}
\value{
An \code{htest} object containing the results of a Kolmogorov-Smirnov test to determine if the paired
probability differences are triangularly distributed. See \link{ks.test} for more details
}
\description{
Assess the plausibility that pairs of nodes are independent by determining if the difference in probability
between the members of the pair are triangularly distributed
}

