% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{plot.probFieldNode}
\alias{plot.probFieldNode}
\title{Plot a node}
\usage{
\method{plot}{probFieldNode}(x, ...)
}
\arguments{
\item{x}{The \code{probFieldNode} of interest}

\item{...}{Additional plot parameters to be added to both the steps and
CDF plots}
}
\description{
Plot the steps and CDF for a \code{probFieldNode}
}

