% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{probabilityOfVariable}
\alias{probabilityOfVariable}
\title{Assessing Variable Probability}
\usage{
probabilityOfVariable(field, randomVariable, node)
}
\arguments{
\item{field}{The \code{probField} to be assessed}

\item{randomVariable}{The random variable value of interest}

\item{node}{The \code{probFieldNode} of interest}
}
\value{
The value of the CDF of the node of interest evaluated at the point where
randomVariable exceeds the observed value in the field
}
\description{
Function to calculate the probability of a random variable being observed at a
given node
}

