% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concepts.R
\name{probField-concept}
\alias{probField-concept}
\title{ProbField - Concept}
\description{
The probability field and inference derived from it are the most important concepts to
understand when using the ogmap package. The probability field is a function which
assigns a probability distribution to any point (node; see \link{nodes-concept}) within its domain.
The probability distributions generated by the field dictate the probability of observing any value
of a random variable of interest at that node.

The probability field is built as a kernel weighted average of observations at a collection of survey nodes.
In essence, when the probability field is asked to assign a probability distribution to some inference node
it determines the distance of the inference node from all survey nodes, then assesses the probability for
the inference node being closest in value to each of the survey nodes. The resultant set of probabilities
are normalized such that the results can be taken as the probability distribution for the new node.
From that probability distribution function, the empirical cummulative distribution function can be calculated
and used to perform any distribution based statistic calculation (see \link{inference-concept}).

The adequacy of the estimates obtained from the field depends critically on the kernel and distance
chosen and the parameters for those functions, see \link{correctness-concept} for more details
on how to ensure satisfactory parameters are chosen.

To fit a probability field (\link{probField})
using the ogmap package, one first needs a \code{data.frame} containing survey nodes (or some
other data structure from which \link{node} can be constructed through the function \link{as.nodes}).
Then a  a kernel function and a distance metric are required (see \link{influence-concept}.
The details are roughly outlined in \link{probField} but for a gentler introduction, see the vigette
"gettingStarted" by running:

\code{vignette("gettingStarted", package = "ogmap")}
}

