% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{probField}
\alias{probField}
\alias{probField.data.frame}
\alias{probField.formula}
\title{Generate a probability field}
\usage{
probField(x, ...)

\method{probField}{formula}(x, data, kernel, distance,
  redundancyCorrected = TRUE, calcCDF = TRUE, breakTies = FALSE, ...)

\method{probField}{data.frame}(x, kernel, distance, response = "y",
  redundancyCorrected = TRUE, calcCDF = TRUE, breakTies = FALSE, ...)
}
\arguments{
\item{x}{Either a \link{formula} or a data.frame of the form described below}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{data}{A data frame containing one column of an observed random variable
and an arbitrary number of predictor variable}

\item{kernel}{A kernel function that generates a weighting given a distance
see \link{kernel} for more information on \code{probFieldKernel}s}

\item{distance}{A function to be applied to two \code{probFieldNode} objects
(produced by \link{node}) to generate the distance between them}

\item{redundancyCorrected}{Logical Whether or not to apply a redundancy correction
to the survey nodes, see \link{calcCDF}}

\item{calcCDF}{Logical whether or not to calculate the step probabilities, CDFs, and redundancy
for each node and attach them to the node objects}

\item{breakTies}{Logical whether or not to add a small number to each random variable
to break ties, see \link{breakTies}}

\item{response}{The name of the column containing the random variable, defaults to "y"
if specified name is not provided will use the first column with a warning}
}
\value{
An object of class \code{probField} containing
\itemize{
\item{\code{survey}}{ An object of class \code{probFieldNodeList} containing the survey nodes
in order of increasing values of their random variable}
\item{\code{kernel}}{ The kernel function used to construct the field}
\item{\code{distance}}{ The distance function used to construct the field}
\item{\code{formula}}{ The formula representing the dependencies amongst variables in the field,
namely, response as a function of the guide variables}
\item{\code{sortOrder}}{ A sorting index to order the original data}
\item{\code{originalOrder}}{ A sorting index to unsort the modified data}
\item{\code{calcCDF}}{ Whether CDFs, steps, and redundancy were calculated and appended to nodes}
\item{\code{redundancyCorrected}}{ Whether the CDFs and steps were corrected for redundancy}
\item{\code{breakTies}}{ Whether ties were broken}
}
}
\description{
Generate a probability field from a data.frame of survey observations
}
\details{
This is the gatekeeper function of the ogmap package, taking a data.frame containing
random variable observations and one or many predictor variables. First the frame is sorted
and converted to a series of nodes. If requested the function will calculate the CDFs and step
probabilities for all the other nodes using kernel weightings and a distance function supplied
by the user. The field represents a base model from which inference regarding the
distribution of potential outcomes given predictor variables, instead of just the expectation
of the distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: 

\item \code{data.frame}: 
}}

