% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistic.R
\name{resample}
\alias{resample}
\alias{resample.probFieldStatistic}
\title{Resample a Statistic}
\usage{
resample(x, ...)

\method{resample}{probFieldStatistic}(x, field, bootSamples, doubleBootSamples,
  ...)
}
\arguments{
\item{x}{A \code{probFieldStatistic}}

\item{...}{Additional parameters (currently only included for the reference in the generic)}

\item{field}{The association \code{probField}}

\item{bootSamples}{The number of bootstrap resamples from the field's survey indices}

\item{doubleBootSamples}{The number of bootstrap resamples to take from each of the sets of
resampled indices above}
}
\value{
A list containing
\itemize{
\item{T1}{ The (numeric) statistic calculated from the field}
\item{T2}{ A vector of length \code{bootSamples} containing resampled statistics}
\item{T3}{ A list of length \code{bootSamples} elements, each element consisting of a vector
of length \code{doubleBootSamples} containing resampled statistics (see details)}
}
}
\description{
Generate a set of bootstrap resampled statistics from a given \code{probField} and an
associated \code{probFieldStatistic}
}
\details{
The resample method relies on a \code{probField}'s ability to generate random survey indices from
the CDFs contained in its \code{probFieldNode}s (see \link{resampledSurveyIndices}).
Since the result function of a \code{probFieldStatistic} is designed to accept survey indices for
which to re-calculate the statistic, a distribution of potential statistics can be created.
The second bootstrap resampling does...
}
\section{Methods (by class)}{
\itemize{
\item \code{probFieldStatistic}: 
}}

