% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{resampledSurveyIndices}
\alias{resampledSurveyIndices}
\alias{resampledSurveyIndices.probField}
\alias{resampledSurveyIndices.probFieldNodeList}
\title{Random sample of survey indices}
\usage{
resampledSurveyIndices(x)

\method{resampledSurveyIndices}{probField}(x)

\method{resampledSurveyIndices}{probFieldNodeList}(x)
}
\arguments{
\item{x}{The \code{probField} or \code{probFieldNodeList} of interest}
}
\value{
A vector of the same length as the \code{probFieldNodeList} (or the survey within the \code{probField})
with random indices generated from the probability distribution at each node
}
\description{
Generate a random sample of survey indices for randomization and monte carlo methods
}
\section{Methods (by class)}{
\itemize{
\item \code{probField}: 

\item \code{probFieldNodeList}: 
}}

