% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{setBandwidths}
\alias{setBandwidths}
\title{Set new bandwidths}
\usage{
setBandwidths(field, bandwidths)
}
\arguments{
\item{field}{The \code{probField} of interest}

\item{bandwidths}{A vector, named vector, or list specifying the new bandwidths}
}
\description{
Set new bandwidths for a probability field and recalculate CDFs, steps, and redundancy
for survey nodes
}

