% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{setMaxCores}
\alias{setMaxCores}
\title{Set maximum number of cores}
\usage{
setMaxCores(n)
}
\arguments{
\item{n}{The maximum number of cores}
}
\value{
invisible null
}
\description{
Set maximum number of cores to use when fitting \code{probfield}s
}

