% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ogmap.R
\docType{data}
\name{stockRecruitment}
\alias{stockRecruitment}
\title{Mock Stock Recruitment Data}
\format{A data frame with 49 rows of 2 variables:
\describe{
  \item{stock}{Adult stock, in 10,000s}
  \item{recruitment}{Juvenile's recruited, in 10,000s}
}}
\description{
A dataset containing the adult stock and juvenile recruitment for
\emph{Imaginarifish examplicus}
}

