% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{subcauchy}
\alias{subcauchy}
\title{Subcauchy Kernel}
\usage{
subcauchy(shape)
}
\arguments{
\item{shape}{The shape parameter governing the subcauchy kernel}
}
\value{
A \code{probFieldKernel} object which wraps a subcauchy kernel function with the shape
exponent \code{shape} and an additional attribute containing the value of \code{shape}
}
\description{
Generate a \code{probFieldKernel} object used to evaluate the relevance of an observation at
a given distance using a subcauchy kernel governed by the parameter shape
}

