% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{subsetField}
\alias{subsetField}
\title{Subset a Field}
\usage{
subsetField(field, indices)
}
\arguments{
\item{field}{The \code{probField} to subset}

\item{indices}{Numeric or logical index to the survey nodes of the probability field}
}
\value{
A \code{probField}
}
\description{
Create a subset of the field
}
\details{
Note that survey nodes are in order of increasing random variable, and not
the original order of the data, indices need to reflect this order
}

