% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceAndKernel.R
\name{trawlDistance}
\alias{trawlDistance}
\title{Distance Between Trawl Surveys}
\usage{
trawlDistance(horizontalScale, verticalScale, latGuide, longGuide,
  verticalGuide, earthRadius = 6367)
}
\arguments{
\item{horizontalScale}{The scaling to be applied to the between lat/long distance}

\item{verticalScale}{The scaling to be applied to the depth distance}

\item{latGuide}{The name of the guide parameter containing the latitudinal coordinate
within the \code{probFieldNode}s}

\item{longGuide}{The name of the guide parameter containing the longitudinal coordinate
within the \code{probFieldNode}s}

\item{verticalGuide}{The name of the guide parameter containing the depth coordinate
within the \code{probFieldNode}s}

\item{earthRadius}{The radius of the earth in kilometers at the approximate latitude of interest
defaults to the radius of the earth at 55N}
}
\description{
Generates a \code{probFieldDistance} object that calculates the scaled distance between two trawl survey nodes
given latitude, longitude, and depth, with respect to independent horizontal and vertical scales.
}
\details{
This function was included for historical reasons and as a sample of a more complex distance function.
Internally, the function generated creates and evaluates both an \link{earthDistance} and a \link{linearDistance}
The \code{earthDistance} evaluates the scaled earth-surface inter-lat/long coordinate distance given the radius
of the earth at your latitude of interest, the default is the earth's radius at 55N. The \link{linearDistance}
evaluates the scaled distance between node depths. The results of evaluating these two functions is summed to
give the overall scaled distance between two nodes. None of the actual computation is performed when this function is
called, instead a function of two nodes is created that will perform all of the computation on demand. See the documentation
for \link{probFieldDistance} for more details.
}

