% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probField.R
\name{variableAtProbability}
\alias{variableAtProbability}
\title{Finding probable value}
\usage{
variableAtProbability(field, probability, node)
}
\arguments{
\item{field}{The \code{probField} of interest}

\item{probability}{The probability the value must not exceed}

\item{node}{The \code{probFieldNode} who's CDF is to be examined}
}
\value{
A numeric value of the random variable
}
\description{
Assess the maximum value expected at a given probability from a field at a specific
node
}

