% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{weightedRelevance}
\alias{weightedRelevance}
\title{Weighted Relevance}
\usage{
weightedRelevance(nodes, surveySource, weights = NULL)
}
\arguments{
\item{nodes}{The \code{probFieldNode} or \code{probFieldNodeList} to for which to find
their influential counterparts}

\item{surveySource}{Either a \code{probField} or \code{probFieldNodeList} from which to find
the indices of relevant member nodes}

\item{weights}{A vector of weights representing the relative importance of influence on
each of the nodes given to the \code{nodes} parameter}
}
\value{
A numeric vector with length equal to the number of nodes in \code{surveySource}
with each element correspondign to the weighted sum of relevance to each of the members of
\code{nodes}
}
\description{
Calculate the weighted relevance of each node in a \code{probField} or
\code{probFieldNodeList} to another group of nodes. The weighted relevance is the sum
of the step probability for each survey node summed accross the other nodes.
}
\details{
surveySource should be a \code{probField} or \code{probFieldNodeList}, although
any object that can be coerced to a \code{probFieldNodeList} works, however doing so
is ill-advised
}

