% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistic.R
\name{weightedSumOfMeans}
\alias{weightedSumOfMeans}
\title{Weighted Sum of Means}
\usage{
weightedSumOfMeans(field, nodes, weights)
}
\arguments{
\item{field}{The \code{probField} of a interest}

\item{nodes}{Ideally a \code{probFieldNodeList}, but generally any list of \code{probFieldNode} objects indicating
where in the domain of \code{field} to generate inference}

\item{weights}{A vector of weights (0-1) for each node in \code{nodes} to weight the statistic}
}
\value{
A \link{probFieldStatistic} object containing the observations from the field, the design results (in this case
a list containing the variance matrix and the weighted step probabilities from the field), and a function obtain and
resample the statistic. See \code{probFieldStatistic} for details of the result function.
}
\description{
Calculated a weighted sum of means for a group of nodes within the domain of a probability field
}

